gcc complaining about type for _isnan.

+++ jsrc/f2.c
@@ -139,7 +139,7 @@ static void jtfmt1(J jt,B e,I m,I d,C*s,I t,C*wv){D y;
    y=*(D*)wv; y=y?y:0.0;  /* -0 to 0 */
    if     (!memcmp(wv,&inf, SZD))strcpy(jt->th2buf,e?"  _" :' '==*s?" _" :"_" );
    else if(!memcmp(wv,&infm,SZD))strcpy(jt->th2buf,e?" __" :' '==*s?" __":"__");
-   else if(_isnan(*wv)          )strcpy(jt->th2buf,e?"  _.":' '==*s?" _.":"_.");
+   else if(_isnan((float) *wv)  )strcpy(jt->th2buf,e?"  _.":' '==*s?" _.":"_.");
    else sprintf(jt->th2buf,s,y);
 }}   /* format one number */
